WAN 2.2 S2V: rename s2v_rope.py → rope_s2v.py for naming consistency
Matches the `<component>_<variant>` pattern used by sibling files —
``transformer_wan_s2v.py``, ``pipeline_wan_s2v.py``,
``test_pipeline_wan_s2v.py``, etc.

Path stays at ``models/transformers/wan2_2/`` — the rope file is S2V-
specific (``rope_precompute`` handles the extended ``noisy + ref +
motion`` sequence) and is imported by both ``motioner`` and
``transformer_wan_s2v``, so it must be a leaf module below both
consumers. Inlining into ``transformer_wan_s2v.py`` would create a
circular import with ``motioner``.
